+++ errors/validation_errors.go
@@ -47,8 +47,9 @@ func NewValidationErrorsBuilder() *ValidationErrorsBuilder {
 }
 
 // Set will add an error on a specific field, if the field already contains an error, it will just add it to the current errors list
-func (v *ValidationErrorsBuilder) Set(field, err string) {
+func (v *ValidationErrorsBuilder) Set(field, err string) *ValidationErrorsBuilder {
 	v.errors[field] = append(v.errors[field], err)
+	return v
 }
 
 // Get will return all errors set for a specific field
@@ -57,12 +58,26 @@ func (v *ValidationErrorsBuilder) Get(field string) []string {
 }
 
 // Merge ValidationErrors with another ValidationErrors
-func (v *ValidationErrorsBuilder) Merge(verr ValidationErrors) {
+func (v *ValidationErrorsBuilder) Merge(verr *ValidationErrors) *ValidationErrorsBuilder {
+	return v.MergeWithPrefix("", verr)
+}
+
+// MergeWithPrefix is merging ValidationErrors in another ValidationError
+// adding a prefix for each error field
+func (v *ValidationErrorsBuilder) MergeWithPrefix(prefix string, verr *ValidationErrors) *ValidationErrorsBuilder {
+	if verr == nil {
+		return v
+	}
+	if prefix != "" && prefix[len(prefix)-1] != '_' {
+		prefix = prefix + "_"
+	}
+
 	for key, values := range verr.Errors {
 		for _, value := range values {
-			v.Set(key, value)
+			v.Set(prefix+key, value)
 		}
 	}
+	return v
 }
 
 // Build will send a ValidationErrors struct if there is some errors or nil if no errors has been defined